requirements.txt
asgiref==3.3.1
boto3==1.18.36
botocore==1.21.36
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
django-autocomplete-light==3.2.10
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.11.1
django-utils-six==2.0
djangorestframework==3.12.1
gunicorn==19.9.0
importlib-metadata==2.0.0
jmespath==0.10.0
Markdown==3.3
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python-dateutil==2.8.2
python-decouple==3.1
pytz==2018.9
s3transfer==0.5.0
six==1.12.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.26.6
zipp==3.3.0
